// finances-api/src/common/types/database.types.ts
export type CategoryType = 'income' | 'expense' | 'investment';
export type TransactionType = 'income' | 'expense';
export type InvestmentType =
  | 'renda_fixa'
  | 'renda_variavel'
  | 'cripto'
  | 'outros';
export type GoalStatus = 'active' | 'completed' | 'cancelled';
export type DebtStatus = 'pending' | 'paid' | 'overdue';